backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.0
python-multipart==0.0.20
python-dotenv==1.0.1

# AI Model Support
google-generativeai==0.8.3
openai==1.59.5
anthropic==0.42.0

# Phase 4: RAG Knowledge Base
chromadb==0.4.22
sentence-transformers==2.3.1
pypdf2==3.0.1
python-docx==1.1.0
markdown==3.5.1

# Phase 4: Export Features
python-pptx==0.6.23

# Utilities
httpx==0.28.1
